+++ CHANGES.md
@@ -1,9 +1,14 @@
-# Changelog
+# Laser Mind Client Release Notes
 
-All notable changes to this project will be documented in this file.
+All notable changes to laser mind client be documented here.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [0.53.0] - 2025-12-23
+
+### Added
+- Scan problem to LPU
 
 ## [0.52.0] - 2025-10-21